Network still not learning.
It is either an issue of vanishing gradients because my weight updates
are quite small or there is a larger issue with my back propagation
algorithm as for some reason layer 1 to layer 2 weight updates are zero.
I tried with just one sample to see if the network could memorise it, it
could not.